Extract bun-test types and add dummy Vitest config
- Move BunTest types to new types.ts to avoid circular deps - Simplify
index exports and remove verbose JSDoc - Make codegen no-op for bun-test
- Add vitest.config.ts and dummy test to satisfy workspace - Adjust
internal imports and utils stub - Add root 'ok' script; remove
native-specific publish scripts
