+++ packages-native/bun-test/package.json
@@ -26,7 +26,7 @@
     "./internal/*": null
   },
   "scripts": {
-    "codegen": "build-utils prepare-v3",
+    "codegen": "echo '@effect-native/bun-test: skip codegen'",
     "build": "pnpm build-esm && pnpm build-annotate && pnpm build-cjs && build-utils pack-v3",
     "build-esm": "tsc -b tsconfig.build.json",
     "build-cjs": "babel build/esm --plugins @babel/transform-export-namespace-from --plugins @babel/transform-modules-commonjs --out-dir build/cjs --source-maps",
