+++ packages-native/bun-test/src/utils.ts
@@ -1,4 +1,7 @@
 /**
  * @since 0.1.0
+ * Intentionally empty to avoid circular exports.
+ * This file exists to keep the path "@effect-native/bun-test/utils" valid
+ * without re-exporting from the main entry.
  */
-export * from "./index.js"
+export {}
