+++ packages-native/bun-test/vitest.dummy.test.ts
@@ -1,0 +1,8 @@
+import { describe, it, expect } from "vitest"
+
+describe("@effect-native/bun-test vitest dummy", () => {
+  it("runs a trivial assertion", () => {
+    expect(1 + 1).toBe(2)
+  })
+})
+
